# cmd/api/docs/swagger.yaml
definitions:
...
  request.Category:
    properties:
      name:
        type: string
    required:
    - name
    type: object
...
  request.Product:
    properties:
      brand:
        type: string
      categoryid:
        type: string
      description:
        type: string
      name:
        type: string
      prize:
        type: integer
      qty_in_stock:
        type: integer
    required:
    - brand
    - categoryid
    - description
    - name
    type: object
...
  respondse.Response:
    properties:
      data: {}
      error: {}
      message:
        type: string
      stastuscode:
        type: integer
    type: object
info:
  contact: {}
paths:
...
  /admin/category/add:
    post:
      consumes:
      - application/json
      description: Admin can create new category from admin panel
      operationId: create-category
      parameters:
      - description: New category name
        in: body
        name: category_name
        required: true
        schema:
          $ref: '#/definitions/request.Category'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/respondse.Response'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/respondse.Response'
      summary: Create new product category
      tags:
      - Product Category
  /admin/category/delete/{category_id}:
    delete:
      consumes:
      - application/json
      description: Admin can delete a category
      operationId: delete-category
      parameters:
      - description: category_id
        in: path
        name: category_id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/respondse.Response'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/respondse.Response'
      summary: Admin can delete a category
      tags:
      - Product Category
  /admin/category/update/{id}:
    patch:
      consumes:
      - application/json
      description: Admin can update category details
      operationId: update-category
      parameters:
      - description: ID of the Category to be updated
        in: path
        name: id
        required: true
        type: string
      - description: category info
        in: body
        name: category_details
        required: true
        schema:
          $ref: '#/definitions/request.Category'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/respondse.Response'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/respondse.Response'
      summary: Admin can update category details
      tags:
      - Product Category
...
  /admin/product/ViewAllProducts:
    get:
      consumes:
      - application/json
      description: users can ses all available products
      operationId: admin-view-all-products
      parameters:
      - description: Page number for pagination
        in: query
        name: page
        type: integer
      - description: Number of items to retrieve per page
        in: query
        name: perPage
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/respondse.Response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/respondse.Response'
      summary: Admins and users can see all available products
      tags:
      - Product
  /admin/product/ViewProduct/{id}:
    get:
      consumes:
      - application/json
      description: Admins and users can see products with product id
      operationId: find-product-by-id
      parameters:
      - description: product id
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/respondse.Response'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/respondse.Response'
      summary: Admins and users can see products with product id
      tags:
      - Product
  /admin/product/delete/{product_id}:
    delete:
      consumes:
      - application/json
      description: Admin can delete a product
      operationId: delete-product
      parameters:
      - description: product_id
        in: path
        name: product_id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/respondse.Response'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/respondse.Response'
      summary: Admin can delete a product
      tags:
      - Product
  /admin/product/save:
    post:
      consumes:
      - application/json
      description: Admins can create new product listings
      operationId: create-product
      parameters:
      - description: new product details
        in: body
        name: new_product_details
        required: true
        schema:
          $ref: '#/definitions/request.Product'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/respondse.Response'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/respondse.Response'
      summary: Admin can create new product listings
      tags:
      - Product
  /admin/product/updateproduct/{id}:
    patch:
      consumes:
      - application/json
      description: Admin can update Product details
      operationId: update-Product
      parameters:
      - description: ID of the product to be updated
        in: path
        name: id
        required: true
        type: string
      - description: Product info
        in: body
        name: category_details
        required: true
        schema:
          $ref: '#/definitions/request.Product'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/respondse.Response'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/respondse.Response'
      summary: Admin can update Product details
      tags:
      - Product
...
  /category/disply/{id}:
    get:
      consumes:
      - application/json
      description: Users and admins can fetch details of a specific category using
        id
      operationId: find-category-by-id
      parameters:
      - description: category id
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/respondse.Response'
        "422":
          description: Unprocessable Entity
          schema:
            $ref: '#/definitions/respondse.Response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/respondse.Response'
      summary: Fetch details of a specific category using category id
      tags:
      - Product Category
  /category/showall:
    get:
      consumes:
      - application/json
      description: Admin, users and unregistered users can see all the available categories
      operationId: view-all-categories
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/respondse.Response'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/respondse.Response'
      summary: View all available categories
      tags:
      - Product Category
...
  /product/AllProducts:
    get:
      consumes:
      - application/json
      description: users can ses all available products
      operationId: user-view-all-products
      parameters:
      - description: Page number for pagination
        in: query
        name: page
        type: integer
      - description: Number of items to retrieve per page
        in: query
        name: perPage
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/respondse.Response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/respondse.Response'
      summary: users can see all available products
      tags:
      - Users
...
swagger: "2.0"